rebase: renamed define.h to macros.h
